regorganize files, make module
* export all functions not starting with '_'
* export all structs
* all code in 'pi.jl'
